Refactor audio processing utilities: implement resampling functions for mulaw and PCM formats, enhance logging for error handling, and reset audio states.
